Ensure session when Single Prod Checkout not set
This adds the ensure_session callback to an order-pay action hook when
Single Product Checkout is disabled.

Previously, in cases where a guest checks out via the order-pay page
(via a payment link sent from the merchant), and Single Product Checkout
is not enabled, no session is set and so the method responsible for
processing once we return from paypal returned early, and guests ended
up at an empty cart page.

This PR fixes that by adding the ensure_session cb to a hook specific
to the order-pay page. We avoid adding this when Single Product
Checkout is active as this cb is added to a more generic hook in this
case.